+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release Notes for Craft Commerce
 
-## Unrelased
+## Unreleased
+
+### Added
+- JSON responses from the `commerce/payments/pay` action now include order information.
 
 ### Changed
 - Purchasable descriptions are only generated based on data from the primary site.
@@ -10,6 +13,7 @@
 - Fixed a bug where donation queries weren’t returning complete results if the primary site had changed.
 - Fixed a PHP error that occurred when converting to a different currency with the `|commerceCurrency` twig filter. ([#1702](https://github.com/craftcms/commerce/issues/1702)) 
 - Fixed a SQL error that occurred when previewing emails on PostgreSQL. ([#1673](https://github.com/craftcms/commerce/issues/1673))
+- Fixed the missing validation of the order condition formula syntax on the discount edit page. ([#1716](https://github.com/craftcms/commerce/issues/1716))
 - Fixed a bug where order history records were ordered incorrectly if they were created at the same time.
 - Fixed a 400 error that could occur when a product type’s Template setting was blank.